replace uselike and useDislike with one single hook
